refactor(storage): extract preuploaded contract to dedicated module
- Add `ensure_matches` method to `PreparedNonDedupBlobUpload` to centralize size/policy validation logic
- Extract `VerifiedPreuploadedNondedupStoreBlob` and its cleanup function into `store/preuploaded_contract.rs`
- Replace inline validation in `VerifiedTempStoreBlob` with `ensure_matches` call
- Add `cleanup_staged_blob_plan_after_contract_failure` in `persist.rs` to handle blob cleanup on contract validation failure
- Update docs to rename `finalize_opaque_upload_session` to `finalize_verified_opaque_upload_session` in upload path tables
